Did some refactoring. Need to fix issues with the Navigation working incorrectly
