+++ src/Header.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import './App.css';
-import Navigation from './Navigation.js';
+
 
 
 
@@ -10,7 +10,7 @@ class Header extends React.Component {
             <>
                 <h1>STYLE BY MISH</h1>
                 <p> STYLE, ART, & CREATIVE DIRECTION</p>
-                <Navigation/>
+                
             </>
         );
     }
