--- src/Home.js
@@ -1,12 +1,0 @@
-import React from 'react';
-
-
-class Home extends React.Component {
-    render() {
-        return (
-            <h1>Welcome to the Home Page</h1>
-        );
-    }
-}
-
-export default Home;
