+++ src/Navigation.js
@@ -1,15 +1,34 @@
-import React from 'react';
 import './App.css';
-import { Link } from 'react-router-dom';
+import { BrowserRouter as Router, Route, Routes, Link } from 'react-router-dom';
 
 
-class Navigation extends React.Component {
-    render() {
-        return (
-            <>
+function Fashion() {
+    return <h2>Fashion</h2>;
+}
+
+function Prop() {
+    return <h2>Prop</h2>;
+}
+
+function Ecomm() {
+    return <h2>E-comm</h2>;
+}
+
+function Contact() {
+    return <h2>Bio/Contact</h2>;
+}
+
+function Clients() {
+    return <h2>Clients</h2>;
+}
+
+function Navigation() {
+
+    return (
+        <Router>
+            <div>
                 <nav>
                     <ul>
-                        <li><Link to="/">HOME</Link></li>
                         <li><Link to="/fashion">FASHION</Link></li>
                         <li><Link to="/prop">PROP</Link></li>
                         <li><Link to="/ecomm">E-COMM</Link></li>
@@ -17,9 +36,17 @@ class Navigation extends React.Component {
                         <li><Link to="/clients">CLIENTS</Link></li>
                     </ul>
                 </nav>
-            </>
-        );
-    }
+                <Routes>
+                    <Route path="/fashion" element={<Fashion />} />
+                    <Route path="/prop" element={<Prop />} />
+                    <Route path="/ecomm" element={<Ecomm />} />
+                    <Route path="/contact" element={<Contact />} />
+                    <Route path="/clients" element={<Clients />} />
+                </Routes>
+            </div>
+        </Router>
+    );
 }
 
 export default Navigation;
+
